+++ Experimental/Dedicated Game Server/Assets/Scripts/Runtime/ConnectionManagement/ConnectionStates/OfflineState.cs
@@ -9,15 +9,15 @@ namespace Unity.Template.Multiplayer.NGO.Runtime.ConnectionManagement
 
         public override void Exit() { }
 
-        public override void StartClient(string ipaddress, int port)
+        public override void StartClient(string ipaddress, ushort port)
         {
-            ConnectionManager.m_ClientConnecting.Configure(ipaddress, (ushort) port);
+            ConnectionManager.m_ClientConnecting.Configure(ipaddress, port);
             ConnectionManager.ChangeState(ConnectionManager.m_ClientConnecting);
         }
 
-        public override void StartServerIP(string ipaddress, int port)
+        public override void StartServerIP(string ipaddress, ushort port)
         {
-            ConnectionManager.m_StartingServer.Configure(ipaddress, (ushort) port);
+            ConnectionManager.m_StartingServer.Configure(ipaddress, port);
             ConnectionManager.ChangeState(ConnectionManager.m_StartingServer);
         }
     }
